Basic Functions
This commit contains two R documents.  One contains code and
documentation to set up the class, and generic getClass methods for my
new class, RegAnalysis.

The other contains code that creates the function fitRegAnalysis, which
creates objects of the type RegAnalysis.  fitRegAnalysis regresses all
combinations of covaraites and intercept on y, as required by the
homework, and returns the matrix of coefficients and a vector of R2’s.

Documentation exists for both of these functions that will eventually
form the help file.