Check for covariate importance
This commit contains two new scripts, RegAnalysis and
performRegAnalysis.

The first creates a new S4 class, RegAnalysis.  This is a subclass of
RegCombin.  It is exactly the same except that it has an extra slot
called MeanR2, which will be filled by the performRegAnalysis function.

The performRegAnalysis calculates the mean R2 value for all models in a
RegCombin object including each variable.  This information is made
into a named vector, which is put into the MeanR2 slot of a RegAnalysis
object, and then returned as a RegAnalysis object.